refactor(sdk/analyze): externalize tool_output_bloat tests
tool_output_bloat.rs was 1676 lines, ~1072 of them the inline test block.
Move the tests verbatim into tool_output_bloat_tests.rs via
`#[cfg(test)] #[path = ...] mod tests;`, matching the patterns_tests.rs /
hotspots_tests.rs convention. Source file is now 606 lines of pure code;
the 29 tests pass identically (paths unchanged). Pure relocation.
